pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyspice"
version = "0.1.0"
description = "Production-quality vertical slice for LTspice orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.0.0",
  "PyLTSpice>=5.4.2",
  "matplotlib>=3.8.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[project.scripts]
pyspice = "pyspice.doctor:main"
